+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.3
+
+- Used the Fisher-Yates shuffle algorithm.
+- Housekeeping.
+
+## 0.1.2
+
+- Added methods and constructors helpful during simulations.
+
 ## 0.1.1
 
 - Housekeeping.